The IIR filter code from before did not work, and a whole bunch of the IIR/other code has been removed from the project. These files still exist, and are still attached and a part of the commits, but no longer in use in the actual project.
The CDMA simulations look like they are working. Some more extensive testing/comparison to the matlab DigiComm code is needed. Plotting output BERs as a function of Eb/N0 is one of the next places to go.

Another minor change is the way modulation is done. Now, modulation is done by copying the reference signals into the waveform array instead of re-calculating every time. hopefully this is faster. In order to do this, the reference array is now a jagged array instead of a 2D array.